ls.gdsn {gdsfmt}	R Documentation

Return the names of child nodes

Description

Get a list of names for its child nodes.

Usage

ls.gdsn(node, include.hidden=FALSE, recursive=FALSE, include.dirs=TRUE)

Arguments

node
an object of class gdsn.class, a GDS node

include.hidden
whether including hidden variables or folders

recursive
whether the listing recurses into directories or not

include.dirs
whether subdirectory names should be included in recursive listings

Value

A vector of characters, or character(0) if node is not a folder.

Author(s)

Xiuwen Zheng

References

http://github.com/zhengxwen/gdsfmt

See Also

cnt.gdsn, objdesp.gdsn, ls.gdsn, index.gdsn

Examples

# cteate a GDS file
f <- createfn.gds("test.gds")

# add a list to "test.gds"
node <- add.gdsn(f, name="list", val=list(x=c(1,2), y=c("T","B","C"), z=TRUE))
ls.gdsn(node)
# "x" "y" "z"

ls.gdsn(f$root)
# "list"

ls.gdsn(f$root, recursive=TRUE)
# "list"   "list/x" "list/y" "list/z"

ls.gdsn(f$root, recursive=TRUE, include.dirs=FALSE)
# "list/x" "list/y" "list/z"

# close the GDS file
closefn.gds(f)

# delete the temporary file
unlink("test.gds", force=TRUE)

[Package gdsfmt version 1.30.0 Index]