print.gds.class {gdsfmt}	R Documentation

Show the information of class "gds.class" and "gdsn.class"

Description

Displays the contents of "gds.class" (a GDS file) and "gdsn.class" (a GDS node).

Usage

## S3 method for class 'gds.class'
print(x, path="", show=TRUE, ...)
## S3 method for class 'gdsn.class'
print(x, expand=TRUE, all=FALSE, nmax=Inf, depth=Inf,
    attribute=FALSE, attribute.trim=FALSE, ...)
## S4 method for signature 'gdsn.class'
show(object)

Arguments

x
an object of class gds.class, a GDS file; or gdsn.class, a GDS node

object
an object of class gds.class, the number of elements in the preview can be specified via the option getOption("gds.preview.num", 6L), while 6L is the default value

path
the path specifying a GDS node with '/' as a separator

show
if TRUE, display the preview of array node

expand
whether enumerate all of child nodes

all
if FALSE, hide GDS nodes with an attribute "R.invisible"

nmax
display nodes within the maximum number nmax

depth
display nodes under maximum depth

attribute
if TRUE, show the attribute(s)

attribute.trim
if TRUE, trim the attribute information if it is too long

...
the arguments passed to or from other methods

Value

None.

Author(s)

Xiuwen Zheng

References

http://github.com/zhengxwen/gdsfmt

Examples

# cteate a GDS file
f <- createfn.gds("test.gds")

add.gdsn(f, "int", 1:100)
add.gdsn(f, "int.matrix", matrix(1:(50*100), nrow=100, ncol=50))
put.attr.gdsn(index.gdsn(f, "int.matrix"), "int", 1:10)

print(f, all=TRUE)
print(f, all=TRUE, attribute=TRUE)
print(f, all=TRUE, attribute=TRUE, attribute.trim=FALSE)

show(index.gdsn(f, "int"))
show(index.gdsn(f, "int.matrix"))

# close the GDS file
closefn.gds(f)

# delete the temporary file
unlink("test.gds", force=TRUE)

[Package gdsfmt version 1.30.0 Index]